+++ gulp-config.js
@@ -82,5 +82,14 @@
       // WebPageTest API key https://www.webpagetest.org/getkey.php
       // key:
     },
+    pa11y: {
+      includeNotices: false,
+      includeWarnings: true,
+      ignore: [
+        'WCAG2AA.Principle2.Guideline2_4.2_4_2.H25.2',
+        'WCAG2AA.Principle2.Guideline2_4.2_4_2.H25.1.NoTitleEl',
+        'WCAG2AA.Principle3.Guideline3_1.3_1_1.H57.2',
+      ]
+    },
   };
 })();
